+++ CHANGELOG.md
@@ -4,6 +4,18 @@ SarPy follows a continuous release process, so there are fairly frequent release
 Since essentially every (squash merge) commit corresponds to a release, specific 
 release points are not being annotated in GitHub.
 
+## [1.3.15] - 2023-02-02
+### Fixed
+- Fixed bug in intersection and bounding box code for geometry elements.
+### Added
+- Pytests for geometry.
+
+## [1.3.14] - 2023-01-11
+### Fixed
+- Fixed deprecated usage of numpy.bool
+### Changed
+- Updated documentation on usage of CPHD consistency check.
+
 ## [1.3.13] - 2022-11-21
 ### Fixed
 - Fixed bug in correctly setting attribution for populate_nitf_information_into_sicd().